refactor: Separate Excel export logic into utility functions for clarity
